Improve cached task context access protection.
The cached next expiring task is now statically stored inside of the task que function.  An updated flag and atomic accessors were added.  The updated flag is now set if an active task’s interval is modified to indicate that the cached task might be invalid and should not be used.  The new method prevents access conflicts for cases where task intervals are modified or tasks are started from contexts other than main.